requirements.txt
numpy==1.23.4
pytest==7.2.0
easydict==1.10
scikit-image==0.19.3
scikit-learn==1.1.3
opencv-python==4.8.0.76
imblearn==0.0
tqdm==4.66.1
pandas==1.5.1
lmdb==1.3.0
tensorboard==2.10.1
torch==1.12.1+cu116 # --index-url https://download.pytorch.org/whl/cu116
torchvision
